requirements.txt
altgraph==0.17
future==0.18.2
mysql-connector-python==8.0.22
pefile==2019.4.18
Pillow==8.0.1
protobuf==3.13.0
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
PyMuPDF==1.18.1
PyPDF2==1.26.0
pywin32==228
pywin32-ctypes==0.2.0
six==1.15.0
